refactor(sse): promote EventStream to public API, deprecate createEventStream
Make the `EventStream` class a first-class public export and move its
implementation out of `utils/internal/` into `utils/event-stream.ts`. The
pure formatters and header helpers stay internal.

`createEventStream` becomes a soft-deprecated compatibility wrapper and moves
to `_deprecated.ts`. Docs, examples, and the migration guide now prefer
`new EventStream(event)`.
