+++ MIGRATION.md
@@ -187,6 +187,7 @@ H3 v2 deprecated some legacy and aliased utilities.
 
 ### Other Utils
 
+- `createEventStream`: Migrate to `new EventStream(event)`.
 - `isStream`: Migrate to `instanceof ReadableStream`.
 - `isWebResponse`: Migrate to `instanceof Response`.
 - `splitCookiesString`: Use `splitSetCookieString` from [cookie-es](https://github.com/unjs/cookie-es).
