+++ docs/1.guide/901.advanced/2.websocket.md
@@ -142,19 +142,22 @@ app.get(
 
 As an alternative to WebSockets, you can use [Server-sent events](https://developer.mozilla.org/en-US/docs/Web/API/Server-sent_events).
 
-H3 has a built-in API to create server-sent events using `createEventStream(event)` utility.
+H3 has a built-in `EventStream` class to create server-sent events. Construct it directly with `new EventStream(event)` and return it from a handler.
+
+> [!NOTE]
+> The `createEventStream(event)` utility is a thin wrapper around `new EventStream(event)` and is kept for compatibility. Prefer the class directly.
 
 ### Example
 
 <!-- automd:file code lang="js" src="../../../examples/server-sent-events.mjs" -->
 
 ```js [server-sent-events.mjs]
-import { H3, serve, createEventStream } from "h3";
+import { H3, serve, EventStream } from "h3";
 
 export const app = new H3();
 
 app.get("/", (event) => {
-  const eventStream = createEventStream(event);
+  const eventStream = new EventStream(event);
 
   // Send a message every second
   const interval = setInterval(async () => {
